Prepare 0.2.2

+++ stochastic_matching/common.py
@@ -209,22 +209,22 @@ def neighbors(i, compressed_incidence):
     Edges of node 0:
 
     >>> neighbors(0, csr_matrix(incidence))
-    array([0, 1, 2], dtype=int32)
+    array([0, 1, 2])
 
     Egde of node 3:
 
     >>> neighbors(3, csr_matrix(incidence))
-    array([2], dtype=int32)
+    array([2])
 
     Nodes of edge 0:
 
     >>> neighbors(0, csc_matrix(incidence))
-    array([0, 1], dtype=int32)
+    array([0, 1])
 
     Nodes of hyperedge 2:
 
     >>> neighbors(2, csc_matrix(incidence))
-    array([0, 1, 2, 3], dtype=int32)
+    array([0, 1, 2, 3])
     """
     return compressed_incidence.indices[compressed_incidence.indptr[i]:compressed_incidence.indptr[i + 1]]
